pyproject.toml
[tool.poetry]
name = "hw-skypro-django-rest-framework"
version = "0.1.0"
description = ""
authors = ["Kovkov-D-S <dmitriy.vavtotrans@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-timezone-field = "^7.1.0"
djangorestframework = "^3.15.2"
dotenv = "^0.9.9"
getenv = "^0.2.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
apiview = "^1.3.25"
coverage = "^7.7.0"
pytz = "^2025.1"
pyzt = "^0.0.2"
drf-yasg = "^1.21.10"
stripe = "^11.6.0"
celery = "^5.4.0"
eventlet = "^0.39.1"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.7.0"
flake8 = "^7.2.0"
gunicorn = "^23.0.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
